% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDAhd.R
\name{runCoDAhd_CountAdd}
\alias{runCoDAhd_CountAdd}
\title{Run various CoDA LR transformations on scRNA-seq data matrix with different count addition}
\usage{
runCoDAhd_CountAdd(
  dat,
  countAdd = "s/gm",
  method = "CLR",
  gene = NULL,
  group = NULL
)
}
\arguments{
\item{dat}{The raw count scRNA-seq matrix (gene(row) x cell(column)).}

\item{countAdd}{The 'count' added to the whole matrix to avoid zero. Options included: s/gm (sum/geometric mean(sum))(default); s/max (sum/maximum(sum)); s/10000(equivalent to LogNorm-CLR); user-defined number(e.g.,1,0.1). s refers to the total count of each cell.}

\item{method}{The CoDA log-ratio method used. Options included: CLR(default); IQLR; LVHA; mdCLR; ILR; manual(should provide gene list as denominator); groupIQLR(should provide group information); groupLVHA(should provide group information).}

\item{gene}{The gene list used as denominator. Should be used with method="manual". Default is NULL.}

\item{group}{The group information used for groupIQLR or groupLVHA. Should be used with method="groupIQLR" or "groupLVHA". Default is NULL.}
}
\value{
The output log-ratio matrix.
}
\description{
This function transforms the raw scRNA-seq matrix to different CoDA log-ratio-based matrix. To avoid zeros, different count addition scheme could be used.
}
\examples{
CLR_dat <- runCoDAhd_CountAdd(raw_dat,countAdd="s/gm",method="CLR");
IQLR1_dat <- runCoDAhd_CountAdd(raw_dat,countAdd=1,method="IQLR");
groupIQLR_dat <- runCoDAhd_CountAdd(raw_dat,countAdd="s/10000",method="groupIQLR",group=metadata$celltype);
HKGLR_dat <- runCoDAhd_CountAdd(raw_dat,countAdd="s/10000",method="manual",gene=c("GAPDH","UBC"));
}
